Simplify x.py by moving functions into scripts
